+++ docs/changelog.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 7.12.2
+
+`2020-12-12`
+
+- fix: resolve no date header width when sticky. [#564](https://github.com/react-component/table/pull/564)
+
+## 7.12.1
+
+`2020-12-08`
+
+- chore: fix build css lost.
+
+## 7.12.0
+
+`2020-12-08`
+
+- docs: use dumi. [#558](https://github.com/react-component/table/pull/558)
+
 ## 7.11.3
 
 `2020-11-25`